chore: separate dev commands into separate file
